Add cleanup of the tmp directories used by verilator (#8)
Each temporary directory created when compiling a `.c`-file with
verilator ends up at about 180MB on my machine. These directories are
placed in `/tmp`, which on my machine is a RAM-disk. I had to resize the
tmpfs, as 8GB was not large enough. I also had to set up a swapfile to
not run out of RAM.

This change makes it so the temporary directory for each test is deleted
once verilator has produced its final executable
`build/xxx/test_xxx.hls`. It is deleted even if the build fails or is
interrupted.

Sidenote: This PR is from a fork, only because I did not have access to
creating branches in this repository.